+++ lib/src/core/len.dart
@@ -3,44 +3,41 @@ import 'rules.dart';
 
 /// Length-related constraints.
 ///
-/// Valid input types: String, num, Iterable, and Map.
+/// Valid input types: [String], [num], [Iterable], and [Map].
 class Len {
-  /// Defines the (exact) length of data a user can enter into a field.
+  /// Defines the (exact) length of data.
   ///
-  /// [len] the length of an input; it must be >= 0.
-  Len(int len, {String error = "length error"})
+  /// [len] the length of the input data; it must be >= 0.
+  /// [error] the error message if the length of the input data != [len].
+  Len(int len, {String? error})
       : assert(len >= 0),
-        _lenVal = _InputLen((int length) => length != len ? error : null);
+        _lenVal =
+            _InputLen((int l) => l != len ? error ?? 'length error' : null);
 
-  /// Defines the minimum length of data a user can enter into a field.
+  /// Defines the shortest length of data.
   ///
-  /// [min] the minimum length of an input; it must be >= 0.
-  /// [short] the error message.
-  Len.min(int min, {String short = 'too short'})
+  /// [min] the shortest length of the input data; it must be >= 0.
+  /// [short] the error message if the length of the input data is < [min].
+  Len.min(int min, {String? short})
       : assert(min >= 0),
-        _lenVal = _InputLen((int length) => length < min ? short : null);
+        _lenVal = _InputLen((int l) => l < min ? short ?? 'too short' : null);
 
-  /// Defines the maximum length of data a user can enter into a field.
+  /// Defines the longest length of data.
   ///
-  /// [max] the maximum length of an input; it must be >= 0.
-  /// [long] the error message.
-  Len.max(int max, {String long = 'too long'})
+  /// [max] the longest length of the input data; it must be >= 0.
+  /// [long] the error message if the length of the input data is > [max].
+  Len.max(int max, {String? long})
       : assert(max >= 0),
-        _lenVal = _InputLen((int length) => length > max ? long : null);
+        _lenVal = _InputLen((int l) => l > max ? long ?? 'too long' : null);
 
-  /// Defines both the minimum and maximum length of data a user can enter into
-  /// a field.
+  /// Defines both the shortest and longest length of the input data.
   ///
   /// [min] the minimum length of an input; it must be >= 0 and <= [max].
-  /// [short] the error message if an input is too short.
   /// [max] the maximum length of an input; it must be >= 0 and >= [min].
-  /// [long] the error message if an input is too long.
-  Len.range({
-    required int min,
-    String short = 'too short',
-    required int max,
-    String long = 'too long',
-  })  : assert(min <= max),
+  /// [short] the error message if the length of the input data is < [min].
+  /// [long] the error message if the length of the input data is > [max].
+  Len.range(int min, int max, {String? short, String? long})
+      : assert(min <= max),
         _lenVal = Rules<Object>([
           Len.min(min, short: short),
           Len.max(max, long: long),
@@ -49,10 +46,10 @@ class Len {
   /// Performs length-related validation.
   final ObjVal _lenVal;
 
-  /// Valid - returns null - if the length of [input] satifies the length
+  /// Valid - returns null - if the length of [input] satisfies the length
   /// requirement.
   ///
-  /// **Note:** the type of [input] must be [String], [Iterable] or [Map].
+  /// **Note:** the type of [input] must be [String], [num],[Iterable] or [Map].
   String? call(Object? input) => _lenVal(input);
 }